Add framestep & code maintenance
・git導入につき編集コメントを削除
・フレームステップ機能の導入
　フラグ追加後ok→gb.cpp:gb::runにて実行VBLANK発生時or終了時に1フレームと考えればおｋ
・右CTRLによる停止処理コード削除。F8でできるしいらないよね。
　同時に重複していたメニュー→ファイル内停止処理も削除
・なんかよくわからないけどgplライセンス表示変更&一部ファイルに追加
　編集年は変更するべきだと思うけどコピーライトは追加しなくてもいい？その辺よくわからない

+++ gb_core/apu.cpp
@@ -398,7 +398,7 @@ inline short apu_snd::wav_produce(int freq,bool interpolation)
 
 	if (freq){
 		if (interpolation){
-			ret=((cur_sample*2500-15000)*wav_cur_pos+(bef_sample*2500-15000)*(0x10000-wav_cur_pos))/0x10000;
+			ret=(short)((cur_sample*2500-15000)*wav_cur_pos+(bef_sample*2500-15000)*(0x10000-wav_cur_pos))/0x10000;
 		}
 		else{
 			ret=cur_sample*2500-15000;
